Rename app.py to resources.py
apps.py defines the topology, resources.py creates Pulumi resources.
